pyproject.toml
[project]
name = "nexrad-pointcloud"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "arm-pyart>=1.18",
    "boto3>=1.34",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
nexrad-fetch = "nexrad_fetch.cli:main"
nexrad-transform = "nexrad_transform.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexrad_fetch", "src/nexrad_transform"]

[tool.ruff]
line-length = 100
